+++ src/test/kotlin/com/exactpro/th2/http/server/TestHTTPResponseNew.kt
@@ -1,4 +1,4 @@
-/*
+package com.exactpro.th2.http.server/*
  * Copyright 2020-2022 Exactpro (Exactpro Systems Limited)
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import rawhttp.core.RawHttp
-import testimpl.TestServerManager
+import com.exactpro.th2.http.server.testimpl.TestServerManager
 
 /*
  * HTTP 1.1
