+++ src/test/kotlin/com/exactpro/th2/http/server/TestHTTPResponseOld.kt
@@ -1,4 +1,4 @@
-/*
+package com.exactpro.th2.http.server/*
  * Copyright 2020-2022 Exactpro (Exactpro Systems Limited)
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,11 +15,14 @@
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
-import testimpl.TestServerManager
+import com.exactpro.th2.http.server.testimpl.TestServerManager
 
-class TestHTTPSResponseOld {
+/*
+ * HTTP 1.0
+ */
+class TestHTTPResponseOld {
     companion object {
-        private val server = TestServerManager(true)
+        private val server = TestServerManager(false)
 
         @BeforeAll
         @JvmStatic
@@ -35,11 +38,12 @@ class TestHTTPSResponseOld {
     }
 
     @Test
-    fun `old http`() {
+    fun responseTest() {
         server.stressSpam("""
             GET / HTTP/1.0
             Host: localhost:${GlobalVariables.PORT}
             User-Agent: client RawHTTP
             """.trimIndent())
     }
+
 }
