+++ src/test/kotlin/com/exactpro/th2/http/server/TestHTTPSResponseOld.kt
@@ -1,4 +1,4 @@
-/*
+package com.exactpro.th2.http.server/*
  * Copyright 2020-2022 Exactpro (Exactpro Systems Limited)
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,9 +15,9 @@
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
-import testimpl.TestServerManager
+import com.exactpro.th2.http.server.testimpl.TestServerManager
 
-class TestHTTPSResponseNew {
+class TestHTTPSResponseOld {
     companion object {
         private val server = TestServerManager(true)
 
@@ -35,13 +35,11 @@ class TestHTTPSResponseNew {
     }
 
     @Test
-    fun responseTest() {
+    fun `old http`() {
         server.stressSpam("""
-            GET / HTTP/1.1
-            Connection: close
+            GET / HTTP/1.0
             Host: localhost:${GlobalVariables.PORT}
             User-Agent: client RawHTTP
             """.trimIndent())
     }
-
 }
